Refactor the common code in sdca_ops.cc out for multiclass SDCA. Change: 139854348
