+++ tensorflow/core/kernels/squared-loss.h
@@ -16,9 +16,6 @@ limitations under the License.
 #ifndef TENSORFLOW_KERNELS_SQUARED_LOSS_H_
 #define TENSORFLOW_KERNELS_SQUARED_LOSS_H_
 
-#include <algorithm>
-#include <cmath>
-
 #include "tensorflow/core/kernels/loss.h"
 
 namespace tensorflow {
